Renamed dummy_plants_and_soil to dummy_plants
