+++ virtual_rainforest/models/animals/animal_communities.py
@@ -16,7 +16,7 @@ from numpy import timedelta64
 from virtual_rainforest.core.logger import LOGGER
 from virtual_rainforest.models.animals.animal_cohorts import AnimalCohort
 from virtual_rainforest.models.animals.decay import CarcassPool, ExcrementPool
-from virtual_rainforest.models.animals.dummy_plants_and_soil import PlantCommunity
+from virtual_rainforest.models.animals.dummy_plants import PlantCommunity
 from virtual_rainforest.models.animals.functional_group import FunctionalGroup
