+++ virtual_rainforest/models/animals/dummy_plants.py
@@ -1,6 +1,5 @@
-"""The ''dummy_plants_and_soil'' classes provides toy plant and soil module
-functionality that are required for setting up and testing the early stages of the
-animal module.
+"""The ''dummy_plants'' classes provides toy plant module functionality that are
+required for setting up and testing the early stages of the animal module.
 """  # noqa: #D205, D415
 
 from __future__ import annotations
